Previously TreeNodePosition now Legacy Position
